requirements.txt
openai>=1.75.0
blaxel>=0.1.0
redis>=5.0.0
redisvl>=0.5.1
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
python-dotenv>=1.0.0
numpy>=1.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
fakeredis>=2.23.0
